requirements.txt
fastapi==0.109.2
uvicorn==0.30.1
pydantic
scikit-learn==1.5.1
numpy
pytest
httpx==0.25.2
starlette==0.36.3
mlflow
sqlalchemy
asyncpg==0.29.0
alembic==1.13.1
pytest-asyncio
python-multipart
aiokafka
redis
prometheus-client
prometheus-fastapi-instrumentator
pyjwt
